@@ -8,9 +8,9 @@ use Symfony\Component\HttpFoundation\Request;
 
 class DefaultController extends AbstractController
 {
-    /**
-     * @Route("/", name="homepage")
-     */
+
+    #[Route('/', name: 'homepage')]
+
     public function indexAction(Request $request)
     {
         // replace this example code with whatever you need
